requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.3.0
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
fastapi==0.136.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.12
Jinja2==3.1.6
MarkupSafe==3.0.3
PyMySQL==1.1.2
nodeenv==1.10.0
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
starlette==1.0.0
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.45.0
